Update VoiceStateUpdateHandler.cs
In case the bot miss the event of the person leaving the chat for reasons such as reconnecting at the time, the next time user joins, it will gain currency for all that time in between. So I made sure that whenever user joins, the previous existingTrack is overwritten and no currency is awarded.